refactor: prepare wave 7 shared readiness
