+++ .github/workflows/ci-shared.yml
@@ -154,6 +154,7 @@ jobs:
 
           printf '%s\n' "${packages}"
           go test ${packages}
+          go -C cmd/e2e test ./... -count=1
 
       - name: Run pinned Go static analysis
         run: scripts/static-analysis.sh
