+++ .textlintignore
@@ -14,3 +14,4 @@ tests/eval/suites/**
 tests/eval/seeds/local_train_100k_v1/**
 tests/eval/snapshots/**
 scripts/e2e-db-cases/**
+cmd/e2e/testdata/**
