+++ architecture/modules/conflict-application.json
@@ -2,7 +2,10 @@
   "schema_version": 1,
   "capability": "conflict-application",
   "source_ownership": [
-    {"path": "cmd/internal/serverapp/conflict_queue_composition.go", "issue": 377}
+    {"path": "cmd/internal/serverapp/conflict_queue_composition.go", "issue": 377},
+    {"path": "cmd/internal/serverapp/conflict_review_worker.go", "issue": 377},
+    {"path": "cmd/internal/serverapp/conflict_review_worker_test.go", "issue": 377},
+    {"path": "internal/repository/evidence_conflict_conflict_integration.e2e", "issue": 377}
   ],
   "completed_issues": [],
   "go": {
@@ -25,5 +28,14 @@
       "reason": "Conflict queue application code still consumes the legacy repository until conflict ownership is isolated."
     }
   ],
-  "workers": []
+  "workers": [
+    {
+      "path": "cmd/internal/serverapp/conflict_review_worker.go",
+      "function": "startConflictReviewWorkers",
+      "kind": "goroutine",
+      "ordinal": 1,
+      "role": "worker",
+      "lifecycle_issue": 381
+    }
+  ]
 }
