+++ architecture/modules/index-retirement.json
@@ -1,0 +1,9 @@
+{
+  "schema_version": 1,
+  "capability": "index-retirement",
+  "completed_issues": [],
+  "go": {"units": []},
+  "browser": {"units": []},
+  "exceptions": [],
+  "workers": []
+}
