+++ architecture/ownership.v1.json
@@ -101,6 +101,7 @@
     "architecture/modules/graph-contract.json",
     "architecture/modules/http-error-policy.json",
     "architecture/modules/http-transport.json",
+    "architecture/modules/index-retirement.json",
     "architecture/modules/mcp-transport.json",
     "architecture/modules/memory-application.json",
     "architecture/modules/migration-composition.json",
@@ -119,6 +120,7 @@
     "architecture/modules/prompt-policy.json",
     "architecture/modules/recall-quality-policy.json",
     "architecture/modules/recall-contract.json",
+    "architecture/modules/recall-postgres-reader.json",
     "architecture/modules/redis-coordination-adapter.json",
     "architecture/modules/remember-application.json",
     "architecture/modules/remember-assessor-port.json",
