+++ internal/domain/read_ids.go
@@ -1,0 +1,22 @@
+package domain
+
+import "strings"
+
+// NormalizeReadIDList trims and de-duplicates caller-provided read IDs while
+// retaining nonblank values for the owning adapter to validate.
+func NormalizeReadIDList(values []string) []string {
+	out := make([]string, 0, len(values))
+	seen := make(map[string]struct{}, len(values))
+	for _, value := range values {
+		value = strings.TrimSpace(value)
+		if value == "" {
+			continue
+		}
+		if _, exists := seen[value]; exists {
+			continue
+		}
+		seen[value] = struct{}{}
+		out = append(out, value)
+	}
+	return out
+}
