+++ scripts/e2e-db-cases/index-retirement.json
@@ -1,0 +1,5 @@
+{
+  "version": 1,
+  "capability": "index-retirement",
+  "cases": []
+}
